Shift some testing code around
